% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMIP5.R
\name{as.array.cmip5data}
\alias{as.array.cmip5data}
\title{Convert a cmip5data object to an array}
\usage{
\method{as.array}{cmip5data}(x, ..., drop = TRUE)
}
\arguments{
\item{x}{A \code{\link{cmip5data}} object}

\item{...}{Other parameters}

\item{drop}{logical. Drop degenerate dimensions?}
}
\value{
The object converted to an array
}
\description{
Convert a cmip5data object to an array
}
\keyword{internal}

