% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkTimePeriod}
\alias{checkTimePeriod}
\title{Check for continuous time periods in CMIP5 files}
\usage{
checkTimePeriod(fileInfo_df)
}
\arguments{
\item{fileInfo_df}{data.frame from getFileInfo}
}
\value{
A data frame showing which ensembles are continuous, and which are not.
In addition to standard identifying fields in the data frame (domain, model,
experiment, variable, and ensemble),
the yrStr field will concatonate the time strings for all ensembles,
allHere is a quick check for yr and mon frequency, start and end dates are
the decimal earliest and latest dates for the ensemble, and files indicates
the number of files in the run.
}
\description{
Check that all time periods match for multi-file ensembles. Before starting to
process what may be hundreds or thousands of CMIP5 files, it's a good idea to verify
that your file set is complete and not missing any years.
Unfortunately it's impossible to automatically check the time signature for
sub-monthly frequencies quickly without opening the netcdf file. These time
signatures will be concatonated and the 'allHere' flag will be returned as
NA for these runs. Note that fixed variables are removed.
}
\details{
Decimal time is (year + (month-1)/12).
}
\note{
this only works for files that are in domains 'fx', 'mon', or 'yr'.
}
\examples{
checkdf <- checkTimePeriod(getFileInfo())
}
\seealso{
\code{\link{getFileInfo}}
}

