% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCMIP5.R
\name{convert_array_to_df}
\alias{convert_array_to_df}
\title{Convert array format cmip5data to data frame format}
\usage{
convert_array_to_df(x, verbose = FALSE)
}
\arguments{
\item{x}{A \code{\link{cmip5data}} object}

\item{verbose}{logical. Print info as we go?}
}
\description{
Convert array format cmip5data to data frame format
}
\details{
Convert array format cmip5data to data frame format, for use with dplyr.
}
\note{
This is an internal RCMIP5 function and not exported.
}
\keyword{internal}

