% Generated by roxygen2 (4.0.1): do not edit by hand
\name{filterDimensionLat}
\alias{filterDimensionLat}
\title{Filter latitude dimension.}
\usage{
filterDimensionLat(x, lats = NULL, verbose = FALSE)
}
\arguments{
\item{x}{cmip5data A \code{\link{cmip5data}} object.}

\item{lats}{numeric vector. Latitudes to filter (in).}

\item{verbose}{logical. Print info as we go?}
}
\value{
A \code{\link{cmip5data}} object.
}
\description{
Filter latitude dimension.
}
\note{
This is an internal RCMIP5 function and not exported.
}
\keyword{internal}

