% Generated by roxygen2 (4.0.1): do not edit by hand
\name{filterDimensionTimeMonths}
\alias{filterDimensionTimeMonths}
\title{Filter time (months) dimension.}
\usage{
filterDimensionTimeMonths(x, months = NULL, verbose = FALSE)
}
\arguments{
\item{x}{cmip5data A \code{\link{cmip5data}} object.}

\item{months}{numeric vector}

\item{verbose}{logical. Print info as we go?}
}
\value{
A \code{\link{cmip5data}} object.
}
\description{
Filter time (months) dimension.
}
\note{
This is an internal RCMIP5 function and not exported.
}
\keyword{internal}

