% Generated by roxygen2 (4.0.1): do not edit by hand
\name{filterDimensionTimeYears}
\alias{filterDimensionTimeYears}
\title{Filter time (years) dimension.}
\usage{
filterDimensionTimeYears(x, years = NULL, verbose = FALSE)
}
\arguments{
\item{x}{cmip5data A \code{\link{cmip5data}} object.}

\item{years}{numeric vector}

\item{verbose}{logical. Print info as we go?}
}
\value{
A \code{\link{cmip5data}} object.
}
\description{
Filter time (years) dimension.
}
\note{
This is an internal RCMIP5 function and not exported.
}
\keyword{internal}

