% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDimensions.R
\name{filterDimensionZ}
\alias{filterDimensionZ}
\title{Filter Z dimension.}
\usage{
filterDimensionZ(x, ZRange = NULL, verbose = FALSE)
}
\arguments{
\item{x}{cmip5data A \code{\link{cmip5data}} object.}

\item{ZRange}{Z values (min, max) to filter data against}

\item{verbose}{logical. Print info as we go?}
}
\value{
A \code{\link{cmip5data}} object.
}
\description{
Filter Z dimension.
}
\note{
This is an internal RCMIP5 function and not exported.
}
\keyword{internal}

