% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getFileInfo}
\alias{getFileInfo}
\title{List all CMIP5 file in a directory tree and parse their filenames for CMIP5
information like experiment, model, and variable names.}
\usage{
getFileInfo(path = ".", recursive = TRUE)
}
\arguments{
\item{path}{root of directory tree}

\item{recursive}{logical. Should the listing recurse into directories?}
}
\value{
data frame containing the following parsed from file names:
        1) full file name, 2) filename without path or '.nc' extension,
        3) variable, 4) domain, 5) model, 6) experiment, 7) ensemble
        8) time period, 9) filesize in KB
}
\description{
List all CMIP5 file in a directory tree and parse their filenames for CMIP5
information like experiment, model, and variable names.
}
\examples{
getFileInfo()
getFileInfo('just_this_dir',recursive=F)
}

