% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadEnsemble}
\alias{loadEnsemble}
\title{Load data for a particular set of experiment/variable/model/ensemble}
\usage{
loadEnsemble(variable = "[^_]+", model = "[^_]+", experiment = "[^_]+",
  ensemble = "[^_]+", domain = "[^_]+", path = ".", recursive = TRUE,
  verbose = TRUE, demo = FALSE)
}
\arguments{
\item{variable}{CMIP5 variable to load}

\item{model}{CMIP5 model to load}

\item{experiment}{CMIP5 experiment to load}

\item{ensemble}{CMIP5 ensemble to load}

\item{domain}{CMIP5 domain to load}

\item{path}{root of directory tree}

\item{recursive}{logical. Recurse into directories?}

\item{verbose}{logical. Print info as we go?}

\item{demo}{logical. Demo mode (reading data from global environment, not disk)?}
}
\value{
A \code{\link{cmip5data}} object.
}
\description{
Load data for a particular set of experiment/variable/model/ensemble
}
\details{
This function is the core of RCMIP5's data-loading. It loads all files matching
the experiment, variable, model, ensemble, and perhaps domain supplied by the caller.
We can also load from the package datasets by specifying DEMO=TRUE.
}

