% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadModel}
\alias{loadModel}
\title{Average all ensemble members of the specified experiment-variable-model combination}
\usage{
loadModel(variable, model, experiment, domain = "[^_]+", path = ".",
  recursive = TRUE, verbose = TRUE, demo = FALSE)
}
\arguments{
\item{variable}{CMIP5 variable to load}

\item{model}{CMIP5 model to load}

\item{experiment}{CMIP5 experiment to load}

\item{domain}{CMIP5 domain to load}

\item{path}{root of directory tree}

\item{recursive}{logical. Should we recurse into directories?}

\item{verbose}{logical. Print info as we go?}

\item{demo}{logical. Demo mode (reading data from global environment, not disk)?}
}
\value{
A \code{\link{cmip5data}} object.
}
\description{
Average all ensemble members of the specified experiment-variable-model combination
}

