% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeAnnualMean}
\alias{makeAnnualMean}
\title{Compute annual mean of a variable}
\usage{
makeAnnualMean(x, yearRange = c(1, Inf), verbose = TRUE, parallel = FALSE,
  FUN = mean)
}
\arguments{
\item{x}{list. A structure returned from loadEnsemble() or loadModel()}

\item{yearRange}{numeric vector. Limit computation to this year range (e.g. for testing)}

\item{verbose}{logical. Print info as we go?}

\item{parallel}{logical. Parallelize if possible?}

\item{FUN}{function. Function to apply across months of year}
}
\value{
list with elements 'files', 'val', 'valUnit', timeUnit', 'calendarStr',
     'lat', 'lon', and 'time'.
}
\description{
Compute annual mean of a variable
}
\examples{
makeAnnualMean(loadModel('nbp','HadGEM2-ES','rcp85',verbose=TRUE))
}

