% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMIP5.R
\name{makePackageData}
\alias{makePackageData}
\title{Make package datasets and write them to disk.}
\usage{
makePackageData(path = "./sampledata", maxSize = Inf, outpath = "./data")
}
\arguments{
\item{path}{root of directory tree}

\item{maxSize}{max size (in MB) of dataset to write}

\item{outpath}{directory to write to}
}
\description{
Make package datasets and write them to disk.
}
\details{
Writes all available ensembles to disk as Rdata files, subject to
a maximum size parameter (CRAN says keep sample data < 5MB).
}
\note{
This is an internal RCMIP5 function and not exported.
}
\keyword{internal}

