% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeExperiments.R
\name{mergeExperiments}
\alias{mergeExperiments}
\title{Merge data for two separate experiments}
\usage{
mergeExperiments(x, y, verbose = FALSE)
}
\arguments{
\item{x}{cmip5data A \code{\link{cmip5data}} object}

\item{y}{cmip5data A \code{\link{cmip5data}} object}

\item{verbose}{logical. Print info as we go?}
}
\value{
A \code{\link{cmip5data}} object.
}
\description{
The most common reason to merge experiments is across time periods--e.g.,
merging 'historical' with one of the RCPs. This function does that, checking
that the merge is appropriate and possible.
}
\details{
The variable, units, spatial grid,
depths/levels, domain, and model must all match. The timesteps must be
identical, and time values non-overlapping. If the time gap between the two experiments
is different than their internal timesteps (e.g., if two monthly data objects are
separated by more than a month) a warning will be printed.
}
\note{
This function is 'in the freezer' (not available to users) for now.

This is an internal RCMIP5 function and not exported.
}
\keyword{internal}

