% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMIP5.R
\name{nvals}
\alias{nvals}
\title{Return number of data values}
\usage{
nvals(x)
}
\arguments{
\item{x}{A \code{\link{cmip5data}} object}
}
\value{
Number of data values in \code{x}.
}
\description{
Return number of data values
}
\details{
Abstracts away getting number of values, the method for which
varies depending on backend implementation.
}
\note{
This is an internal RCMIP5 function and not exported.
}
\keyword{internal}

