% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.summary.cmip5data}
\alias{print.summary.cmip5data}
\title{Print the summary for a 'cmip5data' class object.}
\usage{
\method{print}{summary.cmip5data}(x, ...)
}
\arguments{
\item{x}{A \code{\link{cmip5data}} object}

\item{...}{Other parameters passed to cat}
}
\description{
Print the summary for a 'cmip5data' class object.
}
\details{
Prints a one-line summary of the object
}
\keyword{internal}

