% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMIP5.R
\docType{package}
\name{RCMIP5}
\alias{RCMIP5}
\alias{RCMIP5-package}
\title{Tools for Manipulating and Summarizing CMIP5 Data}
\description{
Working with CMIP5 data can be tricky, forcing scientists to
write custom scripts and programs. The `RCMIP5` package aims
to ease this process, providing a standard, robust, and
high-performance set of functions to (i) explore what data
have been downloaded, (ii) identify missing data, (iii)
average (or apply other mathematical operations) across
experimental ensembles, (iv) produce both temporal and spatial
statistical summaries, and (v) produce
easy-to-work-with graphical and data summaries.
}
\details{
...
}
\references{
Todd-Brown and Bond-Lamberty, 2014: (in prep).

Taylor et al., 2012:
  An overview of CMIP5 and the experiment design, Bulletin of the American
  Meteorological Society, 93, 485-498.
  \url{http://dx.doi.org/10.1175/BAMS-D-11-00094.1}
}
