% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMIP5.R
\name{summary.cmip5data}
\alias{summary.cmip5data}
\title{Summarize a 'cmip5data' class object.}
\usage{
\method{summary}{cmip5data}(object, ...)
}
\arguments{
\item{object}{A \code{\link{cmip5data}} object}

\item{...}{ignored}
}
\value{
A summary structure of the object.
}
\description{
Summarize a 'cmip5data' class object.
}
\details{
Prints a short summary of the object.
}
\keyword{internal}

