% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMIP5.R
\name{vals}
\alias{vals}
\title{Return data values}
\usage{
vals(x)
}
\arguments{
\item{x}{A \code{\link{cmip5data}} object}
}
\value{
Data values in \code{x}.
}
\description{
Return data values
}
\details{
Abstracts away getting values, the method for which
varies depending on backend implementation.
}
\note{
This is an internal RCMIP5 function and not exported.
}
\keyword{internal}

