% Generated by roxygen2 (4.0.1): do not edit by hand
\name{weighted.sum}
\alias{weighted.sum}
\title{Weighted sum--i.e., sum of weighted means. Convenience function}
\usage{
weighted.sum(x, w = rep(1, length(x)), ...)
}
\arguments{
\item{x}{vector of data}

\item{w}{vector of weights}

\item{...}{passed on to weighted.mean}
}
\value{
Weighted mean multipled by sum of weights.
}
\description{
Weighted sum--i.e., sum of weighted means. Convenience function
}
\seealso{
\code{\link{weighted.mean}}
}

