% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.data.frame.cmip5data}
\alias{as.data.frame.cmip5data}
\title{Convert a cmip5data object to a data frame}
\usage{
\method{as.data.frame}{cmip5data}(x, ..., originalNames = FALSE)
}
\arguments{
\item{x}{A \code{\link{cmip5data}} object}

\item{...}{Other parameters}

\item{originalNames}{logical. Use original dimension names from file?}
}
\value{
The object converted to a data frame
}
\description{
Convert a cmip5data object to a data frame
}
\keyword{internal}

