% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGridArea.R
\name{calcGridArea}
\alias{calcGridArea}
\title{Calculate the grid cella area for a centered lat/lon grid}
\usage{
calcGridArea(lon, lat, verbose = FALSE)
}
\arguments{
\item{lon}{longitude coordinates of the grid centers}

\item{lat}{latitude coordinates of the grid centers}

\item{verbose}{logical. Print info as we go?}
}
\value{
The grid cell area in m^2 (meter*meter)
}
\description{
Calculate the grid cella area for a centered lat/lon grid.
}
\details{
Currently the lon must be uniform but the lat does not need to be.
}
\note{
This is an internal RCMIP5 function and not exported.
}
\keyword{internal}

