% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkTimePeriod}
\alias{checkTimePeriod}
\title{Check that all time periods match for multi-file ensembles}
\usage{
checkTimePeriod(fileInfo_df)
}
\arguments{
\item{fileInfo_df}{data.frame from getFileInfo}
}
\value{
data.frame from fileInfo_df: domain, experiment, model, variable, ensemble, and
         yrStr (string - all year boundries in relevent files)
         allHere (boolean - do the year boundries match up?)
         startYr (numeric - decimal time of minimum year
         endYr (numeric - decimal time of maximum year)
}
\description{
Check that all time periods match for multi-file ensembles
}
\details{
Decimal time is (year + (month-1)/12).
}
\examples{
checkTimePeriod(getFileInfo())
}
\seealso{
getFileInfo
}

