% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkTimePeriod}
\alias{checkTimePeriod}
\title{Check for continuous time periods in CMIP5 files}
\usage{
checkTimePeriod(fileInfo_df)
}
\arguments{
\item{fileInfo_df}{data.frame from getFileInfo}
}
\value{
A data frame showing which ensembles are continuous, and which are not.
}
\description{
Check that all time periods match for multi-file ensembles. Before starting to
process what may be hundreds or thousands of CMIP5 files, it's a good idea to verify
that your file set is complete and not missing any years.
}
\details{
Decimal time is (year + (month-1)/12).
}
\note{
this only works for files that are in domains 'fx', '*mon', or '*yr'.
}
\examples{
checkTimePeriod(getFileInfo())
}
\seealso{
\code{\link{getFileInfo}}
}

