% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDimensions.R
\name{filterDimensionLat}
\alias{filterDimensionLat}
\title{Filter latitude dimension.}
\usage{
filterDimensionLat(x, latRange = NULL, verbose = FALSE)
}
\arguments{
\item{x}{cmip5data A \code{\link{cmip5data}} object.}

\item{latRange}{Latitude values (min, max) to filter data against}

\item{verbose}{logical. Print info as we go?}
}
\value{
A \code{\link{cmip5data}} object.
}
\description{
Filter latitude dimension.
}
\note{
This is an internal RCMIP5 function and not exported.
}
\keyword{internal}

