% Generated by roxygen2 (4.0.1): do not edit by hand
\name{filterDimensionLon}
\alias{filterDimensionLon}
\title{Filter longitude dimension.}
\usage{
filterDimensionLon(x, lons = NULL, verbose = FALSE)
}
\arguments{
\item{x}{cmip5data A \code{\link{cmip5data}} object.}

\item{lons}{numeric vector. Longitudes to filter (in).}

\item{verbose}{logical. Print info as we go?}
}
\value{
A \code{\link{cmip5data}} object.
}
\description{
Filter longitude dimension.
}
\note{
This is an internal RCMIP5 function and not exported.
}
\keyword{internal}

