% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDimensions.R
\name{filterDimensionTimeMonths}
\alias{filterDimensionTimeMonths}
\title{Filter time (months) dimension.}
\usage{
filterDimensionTimeMonths(x, monthRange = NULL, verbose = FALSE)
}
\arguments{
\item{x}{cmip5data A \code{\link{cmip5data}} object.}

\item{monthRange}{Months (min, max) to filter data against}

\item{verbose}{logical. Print info as we go?}
}
\value{
A \code{\link{cmip5data}} object.
}
\description{
Filter time (months) dimension.
}
\note{
This is an internal RCMIP5 function and not exported.
}
\keyword{internal}

