% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDimensions.R
\name{filterDimensionTimeYears}
\alias{filterDimensionTimeYears}
\title{Filter time (years) dimension.}
\usage{
filterDimensionTimeYears(x, yearRange = NULL, verbose = FALSE)
}
\arguments{
\item{x}{cmip5data A \code{\link{cmip5data}} object.}

\item{yearRange}{Years (min, max) to filter data against}

\item{verbose}{logical. Print info as we go?}
}
\value{
A \code{\link{cmip5data}} object.
}
\description{
Filter time (years) dimension.
}
\note{
This is an internal RCMIP5 function and not exported.
}
\keyword{internal}

