% Generated by roxygen2 (4.0.1): do not edit by hand
\name{filterDimensionZ}
\alias{filterDimensionZ}
\title{Filter Z dimension.}
\usage{
filterDimensionZ(x, Zs = NULL, verbose = FALSE)
}
\arguments{
\item{x}{cmip5data A \code{\link{cmip5data}} object.}

\item{Zs}{numeric vector. Z values to filter.}

\item{verbose}{logical. Print info as we go?}
}
\value{
A \code{\link{cmip5data}} object.
}
\description{
Filter Z dimension.
}
\note{
This is an internal RCMIP5 function and not exported.
}
\keyword{internal}

