% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadEnsemble}
\alias{loadEnsemble}
\title{Load a unique CMIP5 ensemble}
\usage{
loadEnsemble(variable, model, experiment, ensemble, domain, path = ".",
  recursive = TRUE, verbose = TRUE, demo = FALSE, force.ncdf = FALSE)
}
\arguments{
\item{variable}{CMIP5 variable to load (required)}

\item{model}{CMIP5 model to load (required)}

\item{experiment}{CMIP5 experiment to load (required)}

\item{ensemble}{CMIP5 ensemble to load (required)}

\item{domain}{optinal CMIP5 domain to load (required)}

\item{path}{optional root of directory tree}

\item{recursive}{logical. Recurse into directories?}

\item{verbose}{logical. Print info as we go?}

\item{demo}{logical. Demo mode (reading data from global environment, not disk)?}

\item{force.ncdf}{Force use of the less-desirable ncdf package for testing?}
}
\value{
A \code{\link{cmip5data}} object.
}
\description{
Loads the data for a particular CMIP5 experiment-variable-model-ensemble
combination (one or more files). Returns NULL and a warning if nothing matches.
}
\details{
This function is the core of RCMIP5's data-loading. It loads all files matching
the experiment, variable, model, ensemble, and domain supplied by the caller.
We can also load from the package datasets by specifying DEMO=TRUE.
}
\note{
This is an internal RCMIP5 function and not exported.
}
\keyword{internal}

