% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadEnsemble}
\alias{loadEnsemble}
\title{Load data for a particular set of experiment/variable/model/ensemble}
\usage{
loadEnsemble(variable, model, experiment, ensemble, path = ".",
  recursive = TRUE, verbose = FALSE)
}
\arguments{
\item{variable}{CMIP5 variable to load}

\item{model}{CMIP5 model to load}

\item{experiment}{CMIP5 experiment to load}

\item{ensemble}{CMIP5 ensemble to load}

\item{path}{root of directory tree}

\item{recursive}{logical. Should we recurse into directories?}

\item{verbose}{logical. Print info as we go?}
}
\value{
list with elements 'files', 'val', 'valUnit', timeUnit', 'calendarStr',
     'lat', 'lon', and 'time'. If no files match the requested criteria function
     will return NULL with a warning.
}
\description{
Load data for a particular set of experiment/variable/model/ensemble
}
\examples{
loadEnsemble('nbp','HadGEM2-ES','rcp85','r3i1p1',verbose=T)
}

