% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadModel}
\alias{loadModel}
\title{Average all ensemble members of the specified experiment-variable-model combination}
\usage{
loadModel(variable, model, experiment, path = ".", recursive = TRUE,
  verbose = FALSE, demo = FALSE)
}
\arguments{
\item{variable}{CMIP5 variable of interest}

\item{model}{CMIP5 model of interest}

\item{experiment}{CMIP5 experiment of interest}

\item{path}{root of directory tree}

\item{recursive}{logical. Should we recurse into directories?}

\item{verbose}{logical. Print info as we go?}

\item{demo}{logical. Demo mode (reading data from global environment, not disk)?}
}
\value{
list with elements 'files', 'val', 'valUnit', timeUnit', 'calendarStr',
     'lat', 'lon', and 'time'. If no files match the requested criteria function
     will return NULL with a warning.
}
\description{
Average all ensemble members of the specified experiment-variable-model combination
}
\examples{
loadModel('nbp','HadGEM2-ES','rcp85',verbose=TRUE,demo=TRUE)
}

