% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeAnnualMean}
\alias{makeAnnualMean}
\title{Compute annual mean of a variable}
\usage{
makeAnnualMean(temp.ls, verbose = TRUE)
}
\arguments{
\item{temp.ls}{TODO}

\item{verbose}{logical. Print info as we go?}
}
\value{
TODO
}
\description{
Compute annual mean of a variable
}

