% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeMonthlyMean}
\alias{makeMonthlyMean}
\title{Compute monthly means of a variable}
\usage{
makeMonthlyMean(x, yearRange = c(1, Inf), verbose = TRUE,
  parallel = FALSE, FUN = mean)
}
\arguments{
\item{x}{list. A structure returned from loadEnsemble() or loadModel()}

\item{verbose}{logical. Print info as we go?}

\item{parallel}{logical. Parallelize if possible?}

\item{yearRange}{numeric vector. Limit computation to this year range (e.g. for testing)}

\item{FUN}{function. Function to apply across months of year}
}
\value{
list with elements 'files', 'val', 'valUnit', timeUnit', 'calendarStr',
     'lat', 'lon', and 'time'.
}
\description{
Compute monthly means of a variable
}
\examples{
makeMonthlyMean(loadModel('nbp','HadGEM2-ES','rcp85',verbose=TRUE,demo=TRUE))
}

