% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCMIP5.R
\name{print.cmip5data}
\alias{print.cmip5data}
\title{Print a 'cmip5data' class object.}
\usage{
\method{print}{cmip5data}(x, ...)
}
\arguments{
\item{x}{A \code{\link{cmip5data}} object}

\item{...}{Other parameters passed to cat}
}
\description{
Print a 'cmip5data' class object.
}
\details{
Prints a one-line summary of the object
}
\keyword{internal}

